Implement itchio auth.
The client passes the itch JWT from the ITCHIO_API_KEY environment
variable as the ID field of authenticate_custom. The BeforeAuthItch
hook intercepts this and makes a call to the itch server using this
token. If successful, the hook replaces the ID with the user's ID from
itch. If a custom username was not provided, the username is set to the
itch username.